Factor out Triplers/SearchFilters
